Set up UID in AuthContext to pass UID into STRIPE CHECKOUT
Set up UID in AuthContext to pass UID into STRIPE CHECKOUT

The UID state (user's UID which is MongoDB ID) is now in the global context can be accessed everywhere in the frontend. This means the UID can be passed into the Stripe checkout endpoint that was created by the backend.