Fixed Authentication Bug upon refresh
There was a bug where when the user refreshes the page they could not go to checkout on products because the POST to checkout endpoint requires the user's UID and upon refresh the state of this UID which is in the AuthContext was set to empty string.

This was fixed by altering the AUthContext function and adding some logic in the auth.onAuthStateChanged to make sure upon refresh the user's uid was set globally.

This also meant I had to create a new route and endpoint to fix the bug.